fix(Auth) : 회원 가입 절차 수정
