+++ src/main/java/com/zerobase/wishmarket/domain/authcode/model/dto/AuthCodeResponse.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class AuthCodeResponse {
 
-    private String status;
+    private String message;
 }
