+++ src/main/java/com/zerobase/wishmarket/domain/user/model/dto/SignUpForm.java
@@ -28,4 +28,6 @@ public class SignUpForm {
 
     @NotNull
     private String password;
+
+    private String code;
 }
